// frontend/src/types/reactflow.ts
// 
// Copyright (c) 2024 IB Systems GmbH 
// 
// Licensed under the Apache License, Version 2.0 (the "License"); 
// you may not use this file except in compliance with the License. 
// You may obtain a copy of the License at 
// 
//    http://www.apache.org/licenses/LICENSE-2.0 
// 
// Unless required by applicable law or agreed to in writing, software 
// distributed under the License is distributed on an "AS IS" BASIS, 
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. 
// See the License for the specific language governing permissions and 
// limitations under the License. 
// 


import { Node } from "reactflow";
import { Factory } from "./factory-type";

export interface FlowEditorProps {
  factory: Factory;
  factoryId: string;
}

export interface RelationCounts {
  [key: string]: number;
}

export interface ExtendedNodeData {
  label: string;
  id: string;
  asset_category?: string;
}

export interface Edge {
  hidden?: boolean;
  source: string;
  target: string;
}

export interface ExtendedNode extends Node<ExtendedNodeData> {
  width: number;
  height: number;
  selected: boolean;
  dragging: boolean;
  positionAbsolute: {
    x: number;
    y: number;
  };

  data: {
    type: string;
    label: string;
    id: string;
    class?: string;
    parentId?: string;
  };
  asset_category?: string;
}

export interface FactoryNodeData {
  label?: string;
  type: string;
  undeletable?: boolean;
}